feat: Integrated new format with NepaliDate
