+++ tests/NepaliDate.test.ts
@@ -98,8 +98,8 @@ describe('NepaliDate', () => {
         expect(n.formatNepali('YY-M-D')).toBe('३८-७-१५')
         expect(n.format('YYYY-MM-DD')).toBe('2038-07-15')
         expect(n.format('YY-M-D')).toBe('38-7-15')
-        expect(n.formatNepali('Y-MMMM-ddd')).toBe('२०३८-कार्तिक-शनि')
-        expect(n.format('"YYY" YYYY')).toBe('YYY 2038')
+        expect(n.formatNepali('YYYY-MMMM-ddd')).toBe('२०३८-कार्तिक-शनि')
+        expect(n.format('[YYY] YYYY')).toBe('YYY 2038')
     })
 
     it('checks month, date setting', () => {
